Move proxy to its own package
