+++ SpiNNaker-allocserv/pom.xml
@@ -186,27 +186,6 @@ limitations under the License.
 					</execution>
 				</executions>
 			</plugin>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-jar-plugin</artifactId>
-				<executions>
-					<execution>
-						<phase>package</phase>
-						<goals>
-							<goal>jar</goal>
-						</goals>
-						<configuration>
-							<classifier>proxy</classifier>
-							<includes>**/tcp_proxy/*</includes>
-							<archive>
-								<manifest>
-									<mainClass>uk.ac.manchester.spinnaker.alloc.tcp_proxy.TCPProxy</mainClass>
-								</manifest>
-							</archive>
-						</configuration>
-					</execution>
-				</executions>
-			</plugin>
 		</plugins>
 
 		<pluginManagement>
