+++ SpiNNaker-proxy/pom.xml
@@ -1,0 +1,44 @@
+<!--
+Copyright (c) 2023 The University of Manchester
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    https://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+-->
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>uk.ac.manchester.spinnaker</groupId>
+        <artifactId>SpiNNaker</artifactId>
+        <version>7.1.0-SNAPSHOT</version>
+    </parent>
+    <artifactId>SpiNNaker-proxy</artifactId>
+
+    <description>A simple persistent proxy server.</description>
+    <name>SpiNNaker Proxy Server</name>
+
+    <build>
+        <finalName>spinnaker-proxy</finalName>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-jar-plugin</artifactId>
+                <configuration>
+                    <archive>
+                        <manifest>
+                            <mainClass>uk.ac.manchester.spinnaker.proxy.TCPProxy</mainClass>
+                        </manifest>
+                    </archive>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
+</project>
