+++ pom.xml
@@ -872,6 +872,7 @@ limitations under the License.
 		<module>SpiNNaker-nmpiserv</module>
 		<module>SpiNNaker-nmpimodel</module>
 		<module>SpiNNaker-nmpiexec</module>
+		<module>SpiNNaker-proxy</module>
 	</modules>
 	<name>SpiNNaker Java Host</name>
 	<description>Implementation of the host software for SpiNNaker in Java.</description>
